Refactor test strip warnings into seperate component
